Adds ability to retrieve single job title
