+++ src/main/java/com/detroitlabs/royshowdown/controller/PlayerController.java
@@ -55,6 +55,8 @@ public class PlayerController {
         JobSearchRepository searchRepo = jobService.fetchAllJobs();
         ArrayList<Job> positionTitle = searchRepo.getSearchResult().getSearchResultItems();
 
+        modelMap.put("positionTitle", searchRepo.getSearchResult().getSearchResultItems().get(0).getMatchedObjectDescriptor().getPositionTitle());
+        modelMap.put("searchResultCount", searchRepo.getSearchResult().getSearchResultCount());
         modelMap.put("jobResults", positionTitle);
         modelMap.put("languageCode",searchRepo.getLanguageCode());
         return "testTemplate";
