+++ src/main/java/com/detroitlabs/royshowdown/model/Job.java
@@ -2,31 +2,23 @@ package com.detroitlabs.royshowdown.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonSetter;
 
 import java.util.ArrayList;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Job {
-    private String PositionTitle;
-    private ArrayList<SalaryRange> PositionRemuneration = new ArrayList<SalaryRange>();
 
-    @JsonProperty("PositionTitle")
-    public String getPositionTitle() {
-        return PositionTitle;
+    private JobInfo MatchedObjectDescriptor;
+
+    @JsonProperty("MatchedObjectDescriptor")
+    public JobInfo getMatchedObjectDescriptor() {
+        return MatchedObjectDescriptor;
     }
 
-    @JsonProperty("PositionTitle")
-    public void setPositionTitle(String positionTitle) {
-        PositionTitle = positionTitle;
-    }
 
-    @JsonProperty("PositionRemuneration")
-    public ArrayList<SalaryRange> getPositionRemuneration() {
-        return PositionRemuneration;
-    }
-
-    @JsonProperty("PositionRemuneration")
-    public void setPositionRemuneration(ArrayList<SalaryRange> positionRemuneration) {
-        PositionRemuneration = positionRemuneration;
+    @JsonProperty("MatchedObjectDescriptor")
+    public void setMatchedObjectDescriptor(JobInfo matchedObjectDescriptor) {
+        MatchedObjectDescriptor = matchedObjectDescriptor;
     }
 }
