+++ src/main/java/com/detroitlabs/royshowdown/model/SearchResults.java
@@ -7,8 +7,9 @@ import java.util.ArrayList;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class SearchResults {
-
     private ArrayList<Job> SearchResultItems = new ArrayList<Job>();
+    private int SearchResultCount;
+//    private JobInfo jobInfo;
 
     @JsonProperty("SearchResultItems")
     public ArrayList<Job> getSearchResultItems() {
@@ -19,4 +20,15 @@ public class SearchResults {
     public void setSearchResultItems(ArrayList<Job> searchResultItems) {
         SearchResultItems = searchResultItems;
     }
+
+    @JsonProperty("SearchResultCount")
+    public int getSearchResultCount() {
+        return SearchResultCount;
+    }
+
+    @JsonProperty("SearchResultCount")
+    public void setSearchResultCount(int searchResultCount) {
+        SearchResultCount = searchResultCount;
+    }
+
 }
