+++ src/main/resources/templates/testTemplate.html
@@ -6,9 +6,11 @@
 </head>
 <body>
 <h1>Hello world!</h1>
-<!--<p th:each="Job: ${positionTitle}">Hello!</p>-->
+<p th:each="Job: ${positionTitle}">Hello!</p>
+<p th:text="${positionTitle}"></p>
 <!--<p th:text="${positionTitle.get(0).getPositionTitle()}"></p>-->
 <p th:text="${languageCode}"></p>
+<p th:text="${searchResultCount}"></p>
 
 </body>
 </html>
